+++ src-ui/src/app/components/document-list/bulk-editor/bulk-editor.component.html
@@ -1,6 +1,6 @@
 <div class="row">
   <div class="col-auto mb-2 mb-xl-0" role="group" aria-label="Select">
-    <button class="btn btn-sm btn-outline-danger" (click)="selectNone.next()">
+    <button class="btn btn-sm btn-outline-danger" (click)="documentList.selectNone()">
       <svg width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor">
         <use xlink:href="assets/bootstrap-icons.svg#slash-circle" />
       </svg>
@@ -11,13 +11,13 @@
   <div class="col-auto mb-2 mb-xl-0" role="group" aria-label="Select">
     <label class="mr-2 mb-0">Select:</label>
     <div class="btn-group">
-      <button class="btn btn-sm btn-outline-primary" (click)="selectPage.next()">
+      <button class="btn btn-sm btn-outline-primary" (click)="documentList.selectPage()">
         <svg width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor">
           <use xlink:href="assets/bootstrap-icons.svg#file-earmark-check" />
         </svg>
         Page
       </button>
-      <button class="btn btn-sm btn-outline-primary" (click)="selectAll.next()">
+      <button class="btn btn-sm btn-outline-primary" (click)="documentList.selectAll()">
         <svg width="1em" height="1em" viewBox="0 0 16 16" fill="currentColor">
           <use xlink:href="assets/bootstrap-icons.svg#check-all" />
         </svg>
