+++ src-ui/src/app/components/document-list/document-list.component.html
@@ -79,18 +79,7 @@
 
 <div class="w-100 mb-2 mb-sm-4">
   <app-filter-editor *ngIf="!isBulkEditing" [(filterRules)]="list.filterRules" #filterEditor></app-filter-editor>
-
-  <app-bulk-editor *ngIf="isBulkEditing"
-    [viewDocuments]="list.documents"
-    [(selectedDocuments)]="list.selected"
-    (selectPage)="list.selectPage()"
-    (selectAll)="list.selectAll()"
-    (selectNone)="list.selectNone()"
-    (setTags)="bulkSetTags($event)"
-    (setCorrespondents)="bulkSetCorrespondents($event)"
-    (setDocumentTypes)="bulkSetDocumentTypes($event)"
-    (delete)="bulkDelete()">
-</app-bulk-editor>
+  <app-bulk-editor *ngIf="isBulkEditing"></app-bulk-editor>
 </div>
 
 <div class="d-flex justify-content-between align-items-center">
